refactoring; resolves BV-1038, BV-1037 and BV-1039
